Stop uploading artifacts for each commit
After each release, artifacts are uploaded for the current commit.
When a new commit is pushed, the next cron job creates new artifacts, but these are not linked to the OTA server in order not bother the user with too many (useless OTAs).

So what is the point of uploading these artifacts?
They just waste space.
This commit ends this.

New commits are now only built and released with the next version.

Also add timestamp logging so we can find out where the most of the time is spent in a build.

For example, this release has almost 100 artifacts. 50GB of space 🤯
https://github.com/schnatterer/rooted-graphene/releases/tag/2025020500